feat: (ManageRanksDash.jsx): Dev functionnality for special cases of ranks
The 'Boss' and 'Recrue' ranks are always present and are always the first and last in the list respectively.  The action buttons are grayed out and inactive for these ranks.  If a rank is in second position in the list, its action button to increase is grayed and inactive. And vice versa for the second to last rank in the list.